removed depth texture from post processing buffers and moved fog to forward render pass
