+++ pe-core/src/main/java/org/saar/core/renderer/forward/ForwardRenderingOutput.kt
@@ -1,6 +1,5 @@
 package org.saar.core.renderer.forward
 
-import org.saar.core.postprocessing.PostProcessingBuffers
 import org.saar.core.renderer.RenderingOutput
 import org.saar.core.screen.Screen
 
@@ -10,9 +9,5 @@ class ForwardRenderingOutput(private val screen: Screen,
     override fun to(screen: Screen) = this.screen.copyTo(screen)
 
     override fun toTexture() = this.buffers.albedo
-
-    fun toDepthTexture() = this.buffers.depth
-
-    fun asPostProcessingInput() = PostProcessingBuffers(this.buffers.albedo, this.buffers.depth)
 
 }
